requirements.txt
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
fonttools==4.54.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pillow==11.0.0
PyJWT==2.9.0
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
Werkzeug==3.0.4
